vr_teleop/requirements.txt
pyopenxr>=1.1
opencv-python-headless>=4.8
PyOpenGL>=3.1
glfw>=2.0
numpy>=1.24
